Merged PR 197: Improved movement sway effects for weapon and camera.
Improved movement sway effects for weapon and camera.
There are two types of applied weapon slay, when the player moves and when the player looks/turns.

Added struct to house all related weapon sway configuration.

Refactored how camera movement was done previous. It was using a hack with a camera component to get around UseControllerRotation. I've made it so the Camera is attached to a SpringArm, the SpringArm now uses ControllerRotation which enables the Camera to be able to rotate on it's own for sway effects.

Related work items: #491